requirements.txt
openai>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
browser-use>=0.1.0
selenium>=4.24.0
webdriver-manager>=4.0.2
undetected-chromedriver>=3.5.0


